requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.26.2
netCDF4==1.6.5
python-dateutil==2.8.2
timezonefinder==6.5.2
pytest==8.4.2
tabulate==0.9.0
python-dotenv==1.0.0
skyfield==1.46.0
